requirements.txt
streamlit
langchain
langchain-openai
langgraph
pypdf2
openai
python-dotenv
pydantic
tiktoken